+++ client/components/FilterBox/FilterBox2.jsx
@@ -89,7 +89,7 @@ FilterBox2 = React.createClass({
     }
     return (
     <div className="row">
-      	<div className="col-lg-6" style={{paddingRight:"5px"}}>
+      	<div className="col-lg-6 sm-gutter-right-5px">
         	<div className="filter-box-2 ibox">
         		<div className="ibox-title">
             		<button