+++ client/views/Contact/ContactList.jsx
@@ -2,7 +2,7 @@ ContactList = React.createClass({
 
     showModal(selectedUser) {
         Modal.show({
-            content:<UserProfile item={selectedUser} onChange={this.handleAdd} />
+            content:<UserCard item={selectedUser} onChange={this.handleAdd} />
         })
     },
 
@@ -34,7 +34,8 @@ ContactList = React.createClass({
 
 	render() {
 		var contacts = this.props.items;
-		var canEdit = this.props.onChange!=null;
+		var canCreate = this.props.onChange!=null;
+		var canDelete = false;
 		var component = this;
 		return (
 			<div className="contact-list">
@@ -44,14 +45,14 @@ ContactList = React.createClass({
 			            	className="contact-list-item"
 			                key={idx}
 			            >
-			            	{canEdit?<span className="active-link pull-right" onClick={component.handleRemove.bind(null,idx)}>delete</span>:null}
+			            	{canDelete?<span className="active-link pull-right" onClick={component.handleRemove.bind(null,idx)}>delete</span>:null}
 			            	<div className="active-link" onClick={component.showModal.bind(null,contact)}>
 					            <ContactCard item={contact}/>
 					        </div>
 			            </div>	
 		            )
 			    })}
-			    {canEdit?
+			    {canCreate?
 			    <div 
 			        onClick={component.showModal.bind(null,null)}
 			    >