+++ client/views/Issue/IssueCardDetail.jsx
@@ -1,0 +1,21 @@
+IssueDetail = React.createClass({
+
+    saveItem() {
+        Meteor.call('Issue.save',this.props.item);
+    },
+
+    componentWillMount: function() {
+        this.saveItem = _.debounce(this.saveItem,500);
+    },
+
+    render() {
+        var issue=this.props.item;
+        return (
+            <div className="issue-detail">
+                <IssueSpecArea item={issue} save={this.saveItem} closeCallback={this.props.closeCallback}>
+                    <IssueDynamicArea item={issue} save={this.saveItem} closeCallback={this.props.closeCallback}/>
+                </IssueSpecArea>
+            </div>
+        )
+    }
+})