+++ client/views/Message/NotificationViewSummary.jsx
@@ -1,4 +1,4 @@
-NotificationSummary = React.createClass({
+NotificationViewSummary = React.createClass({
 
     mixins: [ReactMeteorData],
 
@@ -34,29 +34,4 @@ NotificationSummary = React.createClass({
             </div>
 		)
 	}
-})
-
-NotificationView = React.createClass({
-
-    render() {
-        return (
-            <ul className="dropdown-menu dropdown-messages">
-                {
-                this.props.items&&this.props.items.length?
-                	this.props.items.map(function(n){
-                        return (
-                        	<li key={n._id} className="notification-list-item">
-                            	<div className="dropdown-messages-box">
-    	                            <NotificationSummary item={n} />
-                            	</div>
-                        	</li>
-                        )
-                    })
-                :
-                    <li style={{paddingLeft:"10px"}}>No notifications</li>
-                }
-                <li className="browse-button"><a href={FlowRouter.path('messages')}>View all</a></li>
-            </ul>
-        )
-    }
 })