+++ client/views/Team/TeamPageProfile.jsx
@@ -1,21 +1,29 @@
-UserProfilePage = React.createClass({
+TeamProfilePage = React.createClass({
 
     mixins: [ReactMeteorData],
 
     getMeteorData() {
+    	var user, team;
+    	user = Meteor.user();
+    	if(user) {
+    		team = user.getSelectedTeam();
+    	}
 		return {
-			user:Meteor.user()
+			team:team
 		}
 	},
 
 	render() {
+		if(!this.data.team) {
+			return <div/>
+		}
 		return (
 		    <div className="wrapper wrapper-content animated fadeIn">
 		        <div className="row">
 		            <div className="col-lg-6">
-		            	<div className="ibox">
-							<UserCard item={this.data.user}/>
-						</div>
+						<TeamCard
+							item={this.data.team}
+						/>
 					</div>
 				</div>
 			</div>